requirements.txt
pyodbc
python-dotenv
bcrypt
pyjwt